# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing tasks and sandboxes.",
        "title": "Platform Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/auth/sso/callback": {
            "get": {
                "description": "Exchanges the SSO code for a ticket, fetches user info, issues an app JWT, then redirects to the frontend with the token in the URL fragment.",
                "tags": [
                    "auth"
                ],
                "summary": "SSO callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SSO authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "missing code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "SSO check_code or check_user_ticket failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/sso/login": {
            "get": {
                "description": "Redirects the browser to the Didi SSO login page. An optional ?redirect= is forwarded as jumpto.",
                "tags": [
                    "auth"
                ],
                "summary": "SSO login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Frontend redirect target after callback",
                        "name": "redirect",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
...
    },
...
}